# swagger.yaml
openapi: 3.0.3
info:
  title: Smart Vending System API
  description: |
    Flask backend for smart vending: catalog, Omise checkout, members (OTP/JWT), orders, promotions, and admin CRUD under `/api/admin`.
    Interactive docs (Flasgger) are also served by the server at `/apidocs` when the app is running.

    **Socket.IO (same host/port as HTTP, e.g. http://localhost:8000):**
    - Pi agent: auth with `{ "machine_code": "<code>", "token": "<secret>" }` (bcrypt on server). Server emits `job.start` after checkout; agent replies via machine_event RPC (persisted and `job_event_broadcast` to kiosk).
    - Kiosk (machine-ui): auth `{ "role": "kiosk", "kiosk_secret": "..." }` — required in production. Listens for `job_event_broadcast`.
    - Admin dashboard: `{ "role": "admin" }` or `{ "admin_token": "..." }` when `ADMIN_SOCKET_SECRET` is set. Receives `dashboard_update` on machine events.

    Checkout dispatches hardware jobs via Socket.IO `emit_job_start` only (no REST `/api/machines/events`). Manual agent test: POST http://localhost:5000/jobs/start.

    **Docker Compose — default published ports on the host:**

    | Service | Host URL |
    |---------|----------|
    | MySQL (`db`) | `localhost:3307` |
    | API + Socket.IO (`server`) | `http://localhost:8000` |
    | Hardware agent (`client`) | `http://localhost:5000` |
    | machine-ui | `http://localhost:3000` |
    | admin-ui | `http://localhost:3001` |
    | web-ui (member app) | `http://localhost:3002` |
...
servers:
...
paths:
...
  /api/products:
    get:
      tags:
      - Inventory
      summary: List products with stock for a machine
      parameters:
      - name: machine_code
        in: query
        required: false
        schema:
          type: string
        description: Machine identifier (default MP1-001)
      responses:
        '200':
          description: Product rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{product_id}:
    get:
      tags:
      - Inventory
      summary: Get product details by ID
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        '404':
          description: Product not found
...
  /api/members/{phone}:
    get:
      tags:
      - Members
      summary: Lookup member by 10-digit phone (rate limited)
      parameters:
      - name: phone
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Member found
          content:
            application/json:
              schema:
                type: object
                properties:
                  found:
                    type: boolean
                  user_id:
                    type: integer
                  phone_number:
                    type: string
                  display_name:
                    type: string
                    nullable: true
                  points:
                    type: integer
        '400':
          description: Invalid phone
        '404':
          description: Not found
        '429':
          description: Rate limited
    patch:
      tags:
      - Members
      summary: Update member display name (web-ui profile)
      security:
      - bearerAuth: []
      parameters:
      - name: phone
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - display_name
              properties:
                display_name:
                  type: string
      responses:
        '200':
          description: Updated
        '400':
          description: invalid
        '401':
          description: Unauthorized
        '403':
          description: Phone mismatch with token
        '404':
          description: not_found
  /api/members/earn:
    post:
      tags:
      - Members
      summary: Earn points after purchase (rate limited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone_number:
                  type: string
                total_price:
                  type: number
                charge_id:
                  type: string
                  description: Optional link to order
      responses:
        '200':
          description: Points updated
        '400':
          description: Invalid body
        '429':
          description: Rate limited
...
  /api/admin/products:
    get:
      tags:
      - Admin
      summary: List products (paginated, optional q)
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Paginated products
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/PaginatedList'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminProduct'
    post:
      tags:
      - Admin
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - price
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                image_url:
                  type: string
                heating_time:
                  type: integer
                category:
                  type: string
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
...
  /api/admin/products/{product_id}:
    put:
      tags:
      - Admin
      summary: Update product
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                image_url:
                  type: string
                heating_time:
                  type: integer
                category:
                  type: string
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
    delete:
      tags:
      - Admin
      summary: Delete product
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
        '409':
          description: Referenced by slots or orders
...
  /api/members/register:
    post:
      tags:
      - Members
      summary: Register member after OTP (web-ui onboarding)
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Member created
        '409':
          description: already_exists
        '401':
          description: Missing or invalid Bearer token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token (Member JWT or Admin JWT based on endpoint)
  schemas:
    Product:
      type: object
      properties:
        product_id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        heating_time:
          type: integer
          nullable: true
        image_url:
          type: string
          nullable: true
        category:
          type: string
        stock:
          type: integer
...
    PaginatedList:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    AdminProduct:
      type: object
      properties:
        product_id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        image_url:
          type: string
          nullable: true
        heating_time:
          type: integer
          nullable: true
        category:
          type: string
          description: e.g. meat, vegetarian, sweet
...